python/exampleApp/requirements.txt
Pillow==5.1.0
pyscard==1.9.6
